fix(ci): python-coverage-shards line-limit failure
